fix: fallback fonts should not be preloaded
